backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
rasterio==1.3.8
geopandas==0.14.0
shapely==2.0.2
numpy==1.24.3
pandas==2.0.3
pyproj==3.6.0
fiona==1.9.4
Pillow==10.0.0
